refactor: RecentItemRepository 인터페이스화 및 구현체 분리
