+++ README.md
@@ -7,7 +7,7 @@
   - [x] Compose App의 접근성 관련, 기존 Box Clickable을 Button으로 수정
   - [ ] SupervisorJob + Dispatcher.IO의 사용이유 및 관련 코드 리팩토링
   - [ ] viewModelScope.launch {...}의 올바른 사용
-  - [ ] 일관성과 테스트 용이성을 위한 RecentItemRepository의 interface화
+  - [x] 일관성과 테스트 용이성을 위한 RecentItemRepository의 interface화
   - [ ] Room 내 Cart 저장을 통한 API 호출 최소화 및 Local SSOT 설정
   - 
 ## 1단계 구현 기능 사항
