+++ app/src/main/java/woowacourse/shopping/data/repository/RecentItemRepository.kt
@@ -1,40 +1,12 @@
 package woowacourse.shopping.data.repository
 
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
-import woowacourse.shopping.data.localdb.dao.RecentItemDao
-import woowacourse.shopping.data.localdb.mapper.toDomain
-import woowacourse.shopping.data.localdb.mapper.toEntity
 import woowacourse.shopping.model.Product
 
-class RecentItemRepository(
-    private val recentItemDao: RecentItemDao,
-    private val productRepository: ProductRepository,
-) {
-    suspend fun addRecentItem(product: Product) {
-        recentItemDao.insert(product.toEntity(System.currentTimeMillis()))
-        recentItemDao.deleteOldItem()
-    }
+interface RecentItemRepository {
+    suspend fun addRecentItem(product: Product)
 
-    fun getRecentItems(): Flow<List<Product>> =
-        recentItemDao.getRecentItems().map { entities ->
-            entities.mapNotNull { entity ->
-                val product =
-                    runCatching {
-                        productRepository.getProductById(entity.id)
-                    }.getOrNull()
+    fun getRecentItems(): Flow<List<Product>>
 
-                product?.let { entity.toDomain(it) }
-            }
-        }
-
-    suspend fun getLastViewedItem(): Product? {
-        val entity = recentItemDao.getLastViewedItem() ?: return null
-        val product =
-            runCatching {
-                productRepository.getProductById(entity.id)
-            }.getOrNull()
-
-        return product?.let { entity.toDomain(it) }
-    }
+    suspend fun getLastViewedItem(): Product?
 }
